battleship-python/requirements.txt
behave==1.2.6
nose2==0.12.0
pylint==2.16.1
coverage==7.1.0
colorama==0.4.6
opencensus-ext-azure==1.1.8
pytest==7.1.3
